[MF] Fix: current wind was 3.6 times lower than expected

+++ app/src/main/java/wangdaye/com/geometricweather/weather/json/mf/MfCurrentResult.java
@@ -25,8 +25,8 @@ public class MfCurrentResult {
         public Weather weather;
 
         public static class Wind {
-            public double speed;
-            public double gust;
+            public Float speed;
+            public Float gust;
             public Integer direction;
             public String icon;
         }
